requirements.txt
functions-framework==3.*
gspread==5.*
google-auth==2.*
google-cloud-secret-manager==2.*
pandas==2.*

psycopg2-binary==2.9.*